% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load a CSV from the package's extdata folder}
\usage{
load_data(filename)
}
\arguments{
\item{filename}{A character string, the name of the CSV file to load.}
}
\value{
A data frame
}
\description{
Load a CSV from the package's extdata folder
}
